+++ SocketNetworking/Shared/Authentication/AuthenticationProvider.cs
@@ -1,6 +1,5 @@
 ﻿using SocketNetworking.Client;
 using SocketNetworking.Server;
-using SocketNetworking.Shared.PacketSystem;
 using SocketNetworking.Shared.PacketSystem.Packets;
 using SocketNetworking.Shared.Serialization;
 
@@ -76,7 +75,7 @@ namespace SocketNetworking.Shared.Authentication
 
         public int GetLength()
         {
-            return Serialize().Length;
+            return (int)Serialize().Length;
         }
 
         public ByteWriter Serialize()
